config: absolutize state dir, dedupe supervisor resolution, fill wait config
