+++ internal/config/config_test.go
@@ -189,33 +189,3 @@ func TestResolveHTTPToken(t *testing.T) {
 		}
 	})
 }
-
-// TestResolveWaitNeedsNoAgy: ResolveWait must succeed with no agy anywhere on
-// PATH (the wait-only subcommands are pure observers of the job store), while
-// Resolve fails in the same environment, proving the seam matters.
-func TestResolveWaitNeedsNoAgy(t *testing.T) {
-	t.Setenv("AGY_MCP_AGY_PATH", "")
-	t.Setenv("PATH", t.TempDir()) // empty dir: no agy
-	t.Setenv("AGY_MCP_STATE_DIR", "/custom/state")
-
-	if _, err := Resolve(); err == nil {
-		t.Fatal("Resolve succeeded without agy on PATH; the control condition is broken")
-	}
-	c, err := ResolveWait()
-	if err != nil {
-		t.Fatalf("ResolveWait: %v", err)
-	}
-	if c.StateDir != "/custom/state" {
-		t.Fatalf("StateDir = %q, want /custom/state", c.StateDir)
-	}
-	if c.AgyPath != "" {
-		t.Fatalf("wait config resolved a binary it must not need: agy=%q", c.AgyPath)
-	}
-	// SupervisorExe IS resolved (unlike AgyPath): processAlive's fallback
-	// liveness check compares a job's recorded supervisor process name against
-	// it, so leaving it empty would make that comparison always fail and
-	// misreport a live job as dead.
-	if c.SupervisorExe == "" {
-		t.Fatal("SupervisorExe = \"\", want the wait config's own executable path")
-	}
-}
